Llama3ELYZA/requirements.txt
diskcache==5.6.3
Jinja2==3.1.4
keyboard==0.13.5
llama_cpp_python==0.3.1
MarkupSafe==3.0.1
numpy==2.1.2
typing_extensions==4.12.2
